+++ CHANGELOG.md
@@ -27,7 +27,8 @@ for now removed features.
 
 ### Fixed
 
-for any bug fixes.
+- **select**: internal value was not changed when the select is closed and when we used keyboard nav on it (key up, 
+key down... to change the current selected option)
 
 ### Security